# docs/persona-ai-mvp.openapi.yaml
openapi: 3.0.3
info:
  title: TalkTo Persona AI MVP API
  version: 0.1.0
  description: |
    Backend MVP API design for Persona AI.

    v1.0 source priority: TalkTo_개발자용_기능명세서_v1.0.md overrides older Notion rows,
    pasted planning notes, and the legacy MVP baseline where they conflict.

    Current implemented surface: legacy Persona Chat MVP plus Archive-first
    subjects, purpose consents, question cards/interactions, recording upload,
    recording completion, recording list/detail, original recording playback,
    payment products/orders/webhooks/entitlements, and queued paid Archive
    analysis jobs, analysis job status APIs, audited admin retry hooks, and
    admin worker transition endpoints for paid analysis jobs, and paid Memories
    status/search plus segment playback/feedback over indexed analysis segments,
    in-app analysis status notifications, and R&D redacted export opt-in/preview
    guardrails, plus Voice Persona application core routes.

    Guardrails: free Archive upload/completion must not create Preview, STT,
    LLM summary, embedding, memory segment, or analysis job work. Preview/sample
    analysis is deferred/disabled by v1.0. POST
    /api/v1/recordings/{recordingId}/preview-analysis exists only as a disabled
    compatibility endpoint returning 409 feature_deferred and must not enqueue
    Preview, free AI, or analysis job work.

    Not yet implemented: automated provider execution for analysis workers,
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/recordings/{recordingId}/upload-intents/{uploadIntentId}/retry:
    post:
      tags: [Recordings]
      summary: Retry recording upload intent
      description: Creates a fresh upload intent for a failed, canceled, or expired upload intent and returns the new upload URL, storage key, expiry, uploadIntentId, and current recording. Returns retry_not_allowed when the current intent is not retryable.
      parameters:
        - $ref: '#/components/parameters/RecordingId'
        - $ref: '#/components/parameters/UploadIntentId'
      responses:
        '200':
          description: Fresh recording upload intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingUploadIntentResponse'
  /api/v1/recordings/{recordingId}/upload-intents/{uploadIntentId}/cancel:
    post:
      tags: [Recordings]
      summary: Cancel recording upload intent
      description: Cancels an in-progress upload intent with upload_canceled. If the intent is already completed, failed, canceled, or expired, the backend preserves the terminal state and returns the current intent response.
      parameters:
        - $ref: '#/components/parameters/RecordingId'
        - $ref: '#/components/parameters/UploadIntentId'
      responses:
        '200':
          description: Current or cancelled upload intent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingUploadIntentStateResponse'
...
  /api/v1/recordings/{recordingId}/preview-analysis:
    post:
      tags: [Recordings]
      summary: Disabled Preview analysis compatibility endpoint
      description: Exists only for compatibility while Preview/sample analysis is deferred in v1.0. Always returns 409 feature_deferred and does not create Preview, free AI, STT, summary, embedding, memory segment, or analysis job work.
      parameters:
        - $ref: '#/components/parameters/RecordingId'
      responses:
        '409':
          description: Preview analysis is deferred for v1.0.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: feature_deferred
                      message:
                        type: string
                        example: Preview analysis is deferred for v1.0.
                      details:
                        type: object
                        properties:
                          feature:
                            type: string
                            example: preview_analysis
                          retryable:
                            type: boolean
                            example: false
                  meta:
                    type: object
                    properties:
                      timestamp:
                        type: string
                        format: date-time
...
  /api/v1/memory-segments/{memorySegmentId}/playback-url:
    post:
      tags: [Memory Segments]
      summary: Issue a segment-scoped playback URL
      description: Returns a short-lived original-audio playback URL with media-fragment start/end semantics and downloadAllowed=false. Owner-only access is enforced and the sensitive read is audited.
      parameters:
        - in: path
          name: memorySegmentId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemorySegmentPlaybackRequest'
      responses:
        '201':
          description: Segment playback URL issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemorySegmentPlaybackResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/memory-segments/{memorySegmentId}/feedback:
    post:
      tags: [Memory Segments]
      summary: Submit feedback for a Memories segment
      description: Upserts rating/tags/comment and optional reported timestamp corrections for an owned paid Memories result segment.
      parameters:
        - in: path
          name: memorySegmentId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemorySegmentFeedbackRequest'
      responses:
        '201':
          description: Segment feedback saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemorySegmentFeedbackResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/notifications:
    get:
      tags: [Notifications]
      summary: List current user notifications
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - in: query
          name: unreadOnly
          schema:
            type: boolean
      responses:
        '200':
          description: Current user notifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/notifications/{notificationId}/read:
    patch:
      tags: [Notifications]
      summary: Mark a notification as read
      parameters:
        - in: path
          name: notificationId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/persona/sessions:
    post:
      tags: [Persona Runtime]
      summary: Create a gated Voice Persona runtime session
      description: Requires enabled runtime config and Voice Persona consent. This route is separate from legacy /conversations.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePersonaRuntimeSessionRequest'
      responses:
        '201':
          description: Runtime session created
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/persona/sessions/{sessionId}/messages:
    post:
      tags: [Persona Runtime]
      summary: Send a gated Voice Persona runtime message
      description: Uses Memories RAG sources, safety block fallback, TTS playback when configured, and per-session usage accounting.
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePersonaRuntimeMessageRequest'
      responses:
        '201':
          description: Runtime message response
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed or token is missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user lacks the required role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  parameters:
...
    RecordingId:
      in: path
      name: recordingId
      required: true
      schema:
        type: string
        format: uuid
    UploadIntentId:
      in: path
      name: uploadIntentId
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: CONFLICT
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
...
    MemorySegmentPlaybackRequest:
      type: object
      properties:
        bufferMs:
          type: integer
          minimum: 0
          maximum: 30000
          default: 3000
    MemorySegmentPlaybackResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/MemorySegmentPlayback'
    MemorySegmentPlayback:
      type: object
      properties:
        playbackUrl:
          type: string
        expiresAt:
          type: string
          format: date-time
        ttlSeconds:
          type: integer
        downloadAllowed:
          type: boolean
          enum: [false]
        segmentStartMs:
          type: integer
          minimum: 0
        segmentEndMs:
          type: integer
          minimum: 0
        bufferMs:
          type: integer
          minimum: 0
    MemorySegmentFeedbackRequest:
      type: object
      required: [rating]
      properties:
        rating:
          type: string
          enum: [UP, NEUTRAL, DOWN]
        tags:
          type: array
          maxItems: 10
          items:
            type: string
        comment:
          type: string
          maxLength: 500
        reportedStartMs:
          type: integer
          minimum: 0
        reportedEndMs:
          type: integer
          minimum: 0
    MemorySegmentFeedbackResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/MemorySegmentFeedback'
    MemorySegmentFeedback:
      type: object
      properties:
        id:
          type: string
          format: uuid
        memorySegmentId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        rating:
          type: string
          enum: [UP, NEUTRAL, DOWN]
        tags:
          type: array
          items:
            type: string
        comment:
          type: string
          nullable: true
        reportedStartMs:
          type: integer
          nullable: true
        reportedEndMs:
          type: integer
          nullable: true
    NotificationsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
    NotificationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Notification'
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [analysis_completed, analysis_failed, persona_document_status, persona_review_status, persona_build_status]
        title:
          type: string
        body:
          type: string
        resourceType:
          type: string
          nullable: true
        resourceId:
          type: string
          format: uuid
          nullable: true
        subjectId:
          type: string
          format: uuid
          nullable: true
        recordingId:
          type: string
          format: uuid
          nullable: true
        payload:
          type: object
          additionalProperties: true
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    CreatePersonaRuntimeSessionRequest:
      type: object
      required: [applicationId]
      properties:
        applicationId:
          type: string
          format: uuid
    CreatePersonaRuntimeMessageRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          maxLength: 1000
...
    RecordingUploadIntentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/RecordingUploadIntent'
    RecordingUploadIntent:
      type: object
      properties:
        recording:
          $ref: '#/components/schemas/Recording'
        method:
          type: string
          example: PUT
        uploadUrl:
          type: string
          format: uri
        storageKey:
          type: string
        expiresAt:
          type: string
          format: date-time
        uploadIntentId:
          type: string
          format: uuid
        status:
          type: string
          enum: [uploading, completed, failed, canceled, expired]
        source:
          type: string
          enum: [app_recording, share_extension, manual_upload]
        platform:
          type: string
          enum: [ios, android, web]
        singleFile:
          type: boolean
        failureCode:
          type: string
          nullable: true
          enum: [unsupported_mime_type, file_too_large, multi_file_not_supported, upload_expired, object_missing, size_mismatch, checksum_mismatch, upload_canceled, retry_not_allowed, already_completed]
    RecordingUploadIntentStateResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/RecordingUploadIntentState'
        meta:
          type: object
    RecordingUploadIntentState:
      type: object
      properties:
        id:
          type: string
          format: uuid
        recordingId:
          type: string
          format: uuid
        subjectId:
          type: string
          format: uuid
        source:
          type: string
          enum: [app_recording, share_extension, manual_upload]
        platform:
          type: string
          enum: [ios, android, web]
        singleFile:
          type: boolean
        status:
          type: string
          enum: [uploading, completed, failed, canceled, expired]
        failureCode:
          type: string
          nullable: true
          enum: [unsupported_mime_type, file_too_large, multi_file_not_supported, upload_expired, object_missing, size_mismatch, checksum_mismatch, upload_canceled, retry_not_allowed, already_completed]
        originalFilename:
          type: string
        mimeType:
          type: string
        expectedFileSizeBytes:
          type: string
        storageKey:
          type: string
        expiresAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        canceledAt:
          type: string
          format: date-time
          nullable: true
    Recording:
      type: object
      properties:
        id:
          type: string
          format: uuid
        subject_id:
          type: string
          format: uuid
        original_filename:
          type: string
        mime_type:
          type: string
        file_size_bytes:
          type: string
        duration_seconds:
          type: integer
          nullable: true
        upload_status:
          type: string
          enum: [CREATED, UPLOADING, UPLOADED, UPLOAD_FAILED]
        analysis_status:
          type: string
          enum: [NOT_REQUESTED, PROCESSING, COMPLETED, FAILED]
        archive_status:
          type: string
          enum: [pending_upload, active, archived, deletion_requested, deleted]
        analysis_stage:
          type: string
          example: not_analyzed
        memories_status:
          type: string
          example: locked_until_memories
        checksum_status:
          type: string
          enum: [verified, mismatch, not_supported]
        summary:
          nullable: true
        summary_status:
          type: string
          example: locked_until_memories
...